requirements.txt
alembic==1.7.3
blinker==1.4
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.8.6
SQLAlchemy==1.4.25
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.6.0
